fix parameters dtypes

+++ workflows/tf-object-detection-training/utils.py
@@ -85,7 +85,7 @@ def create_pipeline(pipeline_path, model_path, label_path,
         pipeline_config.model.faster_rcnn.first_stage_box_predictor_conv_hyperparams.initializer.truncated_normal_initializer.stddev=float(model_params['first_stage_initializer_stddev'])
         pipeline_config.model.faster_rcnn.first_stage_nms_score_threshold=float(model_params['first_stage_nms_score_threshold'])
         pipeline_config.model.faster_rcnn.first_stage_nms_iou_threshold=float(model_params['first_stage_nms_iou_threshold'])
-        pipeline_config.model.faster_rcnn.first_stage_max_proposals=float(model_params['first_stage_max_proposals'])
+        pipeline_config.model.faster_rcnn.first_stage_max_proposals=int(model_params['first_stage_max_proposals'])
         pipeline_config.model.faster_rcnn.first_stage_localization_loss_weight=float(model_params['first_stage_localization_loss_weight'])
         pipeline_config.model.faster_rcnn.first_stage_objectness_loss_weight=float(model_params['first_stage_objectness_loss_weight'])
         pipeline_config.model.faster_rcnn.initial_crop_size=int(model_params['initial_crop_size'])
@@ -93,13 +93,13 @@ def create_pipeline(pipeline_path, model_path, label_path,
         pipeline_config.model.faster_rcnn.maxpool_stride=int(model_params['maxpool_stride'])        
         pipeline_config.model.faster_rcnn.second_stage_box_predictor.mask_rcnn_box_predictor.fc_hyperparams.regularizer.l2_regularizer.weight=float(model_params['second_stage_regularizer_weight'])
         pipeline_config.model.faster_rcnn.second_stage_box_predictor.mask_rcnn_box_predictor.fc_hyperparams.initializer.variance_scaling_initializer.factor=float(model_params['second_stage_initializer_factor'])
-        pipeline_config.model.faster_rcnn.second_stage_box_predictor.mask_rcnn_box_predictor.fc_hyperparams.initializer.variance_scaling_initializer.mode=float(model_params['second_stage_initializer_mode'])
+        pipeline_config.model.faster_rcnn.second_stage_box_predictor.mask_rcnn_box_predictor.fc_hyperparams.initializer.variance_scaling_initializer.mode=model_params['second_stage_initializer_mode']
         pipeline_config.model.faster_rcnn.second_stage_box_predictor.mask_rcnn_box_predictor.use_dropout=model_params['second_stage_use_dropout']
         pipeline_config.model.faster_rcnn.second_stage_box_predictor.mask_rcnn_box_predictor.dropout_keep_probability=float(model_params['second_stage_dropout_keep_probability'])
         pipeline_config.model.faster_rcnn.second_stage_post_processing.batch_non_max_suppression.score_threshold=float(model_params['second_stage_nms_score_threshold'])
         pipeline_config.model.faster_rcnn.second_stage_post_processing.batch_non_max_suppression.iou_threshold=float(model_params['second_stage_nms_iou_threshold'])
-        pipeline_config.model.faster_rcnn.second_stage_post_processing.batch_non_max_suppression.max_detections_per_class=float(model_params['second_stage_max_detections_per_class'])
-        pipeline_config.model.faster_rcnn.second_stage_post_processing.batch_non_max_suppression.max_total_detections=float(model_params['second_stage_max_detections_max_total_detections'])
+        pipeline_config.model.faster_rcnn.second_stage_post_processing.batch_non_max_suppression.max_detections_per_class=int(model_params['second_stage_max_detections_per_class'])
+        pipeline_config.model.faster_rcnn.second_stage_post_processing.batch_non_max_suppression.max_total_detections=int(model_params['second_stage_max_detections_max_total_detections'])
         pipeline_config.model.faster_rcnn.second_stage_localization_loss_weight=float(model_params['second_stage_localization_loss_weight'])
         pipeline_config.model.faster_rcnn.second_stage_classification_loss_weight=float(model_params['second_stage_classification_loss_weight'])
